+++ genesyscloud/integration/data_source_genesyscloud_integration_test.go
@@ -1,4 +1,4 @@
-package genesyscloud
+package integration
 
 import (
 	"fmt"
