+++ genesyscloud/integration_action/resource_genesyscloud_integration_action.go
@@ -1,4 +1,4 @@
-package genesyscloud
+package integration_action
 
 import (
 	"context"
