+++ genesyscloud/integration_credential/data_source_genesyscloud_integration_credential_test.go
@@ -1,4 +1,4 @@
-package genesyscloud
+package integration_credential
 
 import (
 	"fmt"
