+++ genesyscloud/integration_credential/resource_genesyscloud_integration_credential.go
@@ -1,4 +1,4 @@
-package genesyscloud
+package integration_credential
 
 import (
 	"context"
